+++ resolwe_bio/processes/workflows/rnaseq_variant_calling.py
@@ -21,8 +21,10 @@ class WorkflowRnaseqVariantCalling(Process):
     The pipeline steps include read alignment (STAR), data cleanup
     (MarkDuplicates), splitting reads that contain Ns in their cigar string
     (SplitNCigarReads), base quality recalibration (BaseRecalibrator,
-    ApplyBQSR), variant calling (HaplotypeCaller) and variant filtering
-    (VariantFiltration, SelectVariants).
+    ApplyBQSR), variant calling (HaplotypeCaller), variant filtering
+    (VariantFiltration) and variant annotation (SnpEff). The last step of the
+    pipeline is process Mutations table which prepares variants for ReSDK
+    VariantTables.
     """
 
     slug = "workflow-rnaseq-variantcalling"
@@ -36,7 +38,7 @@ class WorkflowRnaseqVariantCalling(Process):
         },
     }
     data_name = "RNA-seq Variants ({{ reads|name|default('?') }})"
-    version = "1.5.0"
+    version = "2.0.0"
     process_type = "data:workflow:rnaseq:variants"
     category = "Pipeline"
     entity = {
@@ -51,6 +53,11 @@ class WorkflowRnaseqVariantCalling(Process):
             label="Input sample (FASTQ)",
             description="Input data in FASTQ format.",
         )
+        preprocessing = BooleanField(
+            label="Perform reads processing with BBDuk",
+            default=True,
+            description="If your reads have not been processed, set this to True.",
+        )
         ref_seq = DataField(
             data_type="seq:nucleotide", label="Reference FASTA sequence"
         )
@@ -58,11 +65,6 @@ class WorkflowRnaseqVariantCalling(Process):
             data_type="index:star",
             label="Indexed reference genome",
             description="Genome index prepared by STAR aligner indexing tool.",
-        )
-        preprocessing = BooleanField(
-            label="Perform reads processing with BBDuk",
-            default=True,
-            description="If your reads have not been processed, set this to True.",
         )
         dbsnp = DataField(
             data_type="variants:vcf",
@@ -81,48 +83,35 @@ class WorkflowRnaseqVariantCalling(Process):
             "of the genome.",
             required=False,
         )
-        read_group = StringField(
-            label="Replace read groups in BAM",
-            description="Replace read groups in a BAM file. This argument enables the "
-            "user to replace all read groups in the INPUT file with a single new read "
-            "group and assign all reads to this read group in the OUTPUT BAM file. "
-            "Addition or replacement is performed using Picard's AddOrReplaceReadGroups "
-            "tool. Input should take the form of -name=value delimited by a "
-            '";", e.g. "-ID=1;-LB=GENIALIS;-PL=ILLUMINA;-PU=BARCODE;-SM=SAMPLENAME1". '
-            "See tool's documentation for more information on tag names. Note that "
-            "PL, LB, PU and SM are required fields. See caveats of rewriting read groups "
-            "in the documentation.",
-            default="-ID=1;-LB=GENIALIS;-PL=ILLUMINA;-PU=BARCODE;-SM=SAMPLENAME1",
+        clinvar = DataField(
+            data_type="variants:vcf",
+            label="ClinVar VCF file",
+            description="[ClinVar](https://www.ncbi.nlm.nih.gov/clinvar/) is a "
+            "freely available, public archive of human genetic variants and "
+            "interpretations of their significance to disease.",
+            required=False,
         )
-        filter_expressions = ListField(
+        geneset = DataField(
+            data_type="geneset",
+            label="Gene set",
+            description="Select a gene set with genes you are interested in. "
+            "Only variants of genes in the selected gene set will be in the "
+            "output.",
+            required=False,
+            disabled="mutations",
+        )
+        mutations = ListField(
             StringField(),
-            label="Expressions used with INFO fields to filter",
-            description="VariantFiltration accepts any number of JEXL expressions "
-            "(so you can have two named filters by using --filter-name One "
-            "--filter-expression 'X < 1' --filter-name Two --filter-expression 'X > 2'). "
-            "It is preferable to use multiple expressions, each specifying an individual "
-            "filter criteria, to a single compound expression that specifies multiple "
-            "filter criteria. Input expressions one by one and press ENTER after each "
-            "expression. Examples of filter expression: 'FS > 30', 'DP > 10'.",
-            default=["FS > 30.0", "QD < 2.0", "DP < 10.0"],
-        )
-        filter_name = ListField(
-            StringField(),
-            label="Names to use for the list of filters",
-            description="This name is put in the FILTER field for variants that get "
-            "filtered. Note that there must be a 1-to-1 mapping between filter expressions "
-            "and filter names. Input expressions one by one and press ENTER after each name. "
-            "Warning: filter names should be in the same order as filter expressions. "
-            "Example: you specified filter expressions 'FS > 30' and 'DP > 10', now "
-            "specify filter names 'FS' and 'DP'.",
-            default=["FS", "QD", "DP"],
-        )
-        exclude_filtered = BooleanField(
-            label="Exclude filtered sites",
-            default=True,
-            description="If this flag is enabled, sites that have been marked as filtered "
-            "(i.e. have anything other than `.` or `PASS` in the FILTER field) will be excluded "
-            "from the output.",
+            label="Gene and its mutations",
+            description="Insert the gene you are interested in, together "
+            "with mutations. First enter the name of the gene and then "
+            "the mutations. Seperate gene from mutations with ':' and mutations "
+            "with ','. Example of an input: 'KRAS: Gly12, Gly61'. Press enter "
+            "after each input (gene + mutations). NOTE: Field only accepts "
+            "three character amino acid symbols. If you use this option, "
+            "the selected geneset will not be used for Mutations table process.",
+            required=False,
+            disabled="geneset",
         )
 
         class Bbduk:
@@ -207,6 +196,23 @@ class WorkflowRnaseqVariantCalling(Process):
                 description="Output of unmapped reads in the SAM format.",
             )
 
+        class BAMProcessing:
+            """BAM file processing options."""
+
+            read_group = StringField(
+                label="Replace read groups in BAM",
+                description="Replace read groups in a BAM file. This argument enables the "
+                "user to replace all read groups in the INPUT file with a single new read "
+                "group and assign all reads to this read group in the OUTPUT BAM file. "
+                "Addition or replacement is performed using Picard's AddOrReplaceReadGroups "
+                "tool. Input should take the form of -name=value delimited by a "
+                '";", e.g. "-ID=1;-LB=GENIALIS;-PL=ILLUMINA;-PU=BARCODE;-SM=SAMPLENAME1". '
+                "See tool's documentation for more information on tag names. Note that "
+                "PL, LB, PU and SM are required fields. See caveats of rewriting read groups "
+                "in the documentation.",
+                default="-ID=1;-LB=GENIALIS;-PL=ILLUMINA;-PU=BARCODE;-SM=SAMPLENAME1",
+            )
+
         class HaplotypeCaller:
             """GATK HaplotypeCaller options."""
 
@@ -221,18 +227,124 @@ class WorkflowRnaseqVariantCalling(Process):
                 default=True,
                 description="Suitable option for RNA-seq variant calling.",
             )
+            interval_padding = IntegerField(
+                label="Interval padding",
+                default=100,
+                description="Amount of padding (in bp) to add to each interval "
+                "you are including. The recommended value is 100. Set to 0 if you "
+                "want to turn it off.",
+                hidden="!intervals",
+            )
 
-        class SelectVariants:
-            """GATK SelectVariants options."""
+        class VariantFiltration:
+            """GATK VariantFiltration options."""
 
-            select_type = ListField(
+            filter_expressions = ListField(
                 StringField(),
-                label="Select only a certain type of variants from the input file",
-                description="This argument selects particular kinds of variants out of a list. If "
-                "left empty, there is no type selection and all variant types are considered for "
-                "other selection criteria. Valid types are INDEL, SNP, MIXED, MNP, SYMBOLIC, "
-                "NO_VARIATION. Can be specified multiple times.",
-                default=["SNP"],
+                label="Expressions used with INFO fields to filter",
+                description="VariantFiltration accepts any number of JEXL expressions "
+                "(so you can have two named filters by using --filter-name One "
+                "--filter-expression 'X < 1' --filter-name Two --filter-expression 'X > 2'). "
+                "It is preferable to use multiple expressions, each specifying an individual "
+                "filter criteria, to a single compound expression that specifies multiple "
+                "filter criteria. Input expressions one by one and press ENTER after each "
+                "expression. Examples of filter expression: 'FS > 30', 'DP > 10'.",
+                default=["FS > 30.0", "QD < 2.0", "DP < 10.0"],
+            )
+            filter_name = ListField(
+                StringField(),
+                label="Names to use for the list of filters",
+                description="This name is put in the FILTER field for variants that get "
+                "filtered. Note that there must be a 1-to-1 mapping between filter expressions "
+                "and filter names. Input expressions one by one and press ENTER after each name. "
+                "Warning: filter names should be in the same order as filter expressions. "
+                "Example: you specified filter expressions 'FS > 30' and 'DP > 10', now "
+                "specify filter names 'FS' and 'DP'.",
+                default=["FS", "QD", "DP"],
+            )
+
+        class SnpEff:
+            """SnpEff options."""
+
+            filtering_options = StringField(
+                label="SnpEff filtering expressions",
+                description="Filter annotated VCF file using arbitraty expressions."
+                "Examples of filtering expressions: '(ANN[*].GENE = 'PSD3')' "
+                "or '( REF = 'A' )' or "
+                "'(countHom() > 3) | (( exists INDEL ) & (QUAL >= 20)) | (QUAL >= 30 )'."
+                "For more information checkout the official documentation of [SnpSift]"
+                "(https://pcingola.github.io/SnpEff/ss_filter/)",
+                required=False,
+            )
+
+        class MutationsTable:
+            """Mutations table options."""
+
+            vcf_fields = ListField(
+                StringField(),
+                label="Select VCF fields",
+                description="The name of a standard VCF field or an "
+                "INFO field to include in the output table. "
+                "The field can be any standard VCF column (e.g. CHROM, ID, QUAL) "
+                "or any annotation name in the INFO field (e.g. AC, AF). "
+                "Required fields are CHROM, POS, ID, REF and ANN. If your variants "
+                "file was annotated with clinvar information then fields CLNDN, "
+                "CLNSIG and CLNSIGCONF might be of your interest.",
+                default=[
+                    "CHROM",
+                    "POS",
+                    "ID",
+                    "QUAL",
+                    "REF",
+                    "ALT",
+                    "DP",
+                    "FILTER",
+                    "ANN",
+                    "CLNDN",
+                    "CLNSIG",
+                ],
+            )
+            ann_fields = ListField(
+                StringField(),
+                label="ANN fields to use",
+                description="Only use specific fields from the SnpEff ANN "
+                "field. All available fields: Allele | Annotation | Annotation_Impact "
+                "| Gene_Name | Gene_ID | Feature_Type | Feature_ID | Transcript_BioType "
+                "| Rank | HGVS.c | HGVS.p | cDNA.pos / cDNA.length | CDS.pos / CDS.length "
+                "| AA.pos / AA.length | Distance | ERRORS / WARNINGS / INFO' ."
+                "Fields are seperated by '|'. For more information, follow this [link]"
+                "(https://pcingola.github.io/SnpEff/se_inputoutput/#ann-field-vcf-output-files).",
+                default=[
+                    "Allele",
+                    "Annotation",
+                    "Annotation_Impact",
+                    "Gene_Name",
+                    "Feature_ID",
+                    "HGVS.p",
+                ],
+            )
+            split_alleles = BooleanField(
+                label="Split multi-allelic records into multiple lines",
+                description="By default, a variant record with multiple "
+                "ALT alleles will be summarized in one line, with per "
+                "alt-allele fields (e.g. allele depth) separated by commas."
+                "This may cause difficulty when the table is loaded by "
+                "an R script, for example. Use this flag to write multi-allelic "
+                "records on separate lines of output.",
+                default=True,
+            )
+            show_filtered = BooleanField(
+                label="Include filtered records in the output",
+                default=True,
+                description="Include filtered records in the output of the GATK "
+                "VariantsToTable.",
+            )
+            gf_fields = ListField(
+                StringField(),
+                label="Include FORMAT/sample-level fields",
+                default=[
+                    "GT",
+                ],
             )
 
         class Advanced:
@@ -257,13 +369,16 @@ class WorkflowRnaseqVariantCalling(Process):
             Alignment,
             label="Alignment with STAR",
         )
+        bam_processing = GroupField(BAMProcessing, label="Processing of BAM file")
         haplotype_caller = GroupField(
             HaplotypeCaller, label="Options for HaplotypeCaller"
         )
-        select_variants = GroupField(
-            SelectVariants,
-            label="Options for GATK SelectVariants",
-            hidden="!exclude_filtered",
+        variant_filtration = GroupField(
+            VariantFiltration, label="Options for GATK VariantFiltration"
+        )
+        snpeff = GroupField(SnpEff, label="SnpEff options")
+        mutations_table = GroupField(
+            MutationsTable, label="Options for Mutations table"
         )
         advanced = GroupField(Advanced, label="Advanced options")
 
@@ -325,6 +440,21 @@ class WorkflowRnaseqVariantCalling(Process):
                         f"File with INDELs is based on {indel.output.build}, "
                         f"while STAR index is based on {inputs.genome.output.build}."
                     )
+
+        if not inputs.mutations and not inputs.geneset:
+            self.error(
+                "Mutations or geneset were not specified. You must either enter desired "
+                "mutations or select your geneset of interest."
+            )
+
+        if not all(
+            field in inputs.mutations_table.vcf_fields
+            for field in ["CHROM", "POS", "ID", "REF", "ANN"]
+        ):
+            self.error(
+                "Input VCF fields do not contain all required values. "
+                "Required fields are CHROM, POS, ID, REF and ANN."
+            )
 
         if inputs.preprocessing:
 
@@ -390,55 +520,24 @@ class WorkflowRnaseqVariantCalling(Process):
             name=f"Aligned ({inputs.reads.name})",
         )
 
-        mark_duplicates = Data.create(
-            process=BioProcess.get_latest(slug="markduplicates"),
-            input={
-                "bam": alignment,
-                "advanced": {
-                    "java_gc_threads": inputs.advanced.java_gc_threads,
-                    "max_heap_size": inputs.advanced.max_heap_size,
-                },
-            },
-            name=f"Marked duplicates ({inputs.reads.name})",
-        )
-
-        split_ncigar = Data.create(
-            process=BioProcess.get_latest(slug="gatk-split-ncigar"),
-            input={
-                "bam": mark_duplicates,
-                "ref_seq": inputs.ref_seq,
-                "advanced": {
-                    "java_gc_threads": inputs.advanced.java_gc_threads,
-                    "max_heap_size": inputs.advanced.max_heap_size,
-                },
-            },
-            name=f"Split reads ({inputs.reads.name})",
-        )
-
-        input_bqsr = {
-            "bam": split_ncigar,
-            "reference": inputs.ref_seq,
+        input_preprocess = {
+            "bam": alignment,
+            "ref_seq": inputs.ref_seq,
             "known_sites": [inputs.dbsnp],
-            "read_group": inputs.read_group,
-            "advanced": {
-                "use_original_qualities": True,
-                "java_gc_threads": inputs.advanced.java_gc_threads,
-                "max_heap_size": inputs.advanced.max_heap_size,
-            },
+            "read_group": inputs.bam_processing.read_group,
         }
-
         if inputs.indels:
             for indel in inputs.indels:
-                input_bqsr["known_sites"].append(indel)
+                input_preprocess["known_sites"].append(indel)
 
-        bqsr = Data.create(
-            process=BioProcess.get_latest(slug="bqsr"),
-            input=input_bqsr,
-            name=f"Recalibrated ({inputs.reads.name})",
+        preprocess = Data.create(
+            process=BioProcess.get_latest(slug="rnaseq-vc-preprocess"),
+            input=input_preprocess,
+            name=f"Preprocessed ({inputs.reads.name})",
         )
 
         input_hc = {
-            "alignment": bqsr,
+            "alignment": preprocess,
             "genome": inputs.ref_seq,
             "dbsnp": inputs.dbsnp,
             "stand_call_conf": inputs.haplotype_caller.stand_call_conf,
@@ -452,6 +551,9 @@ class WorkflowRnaseqVariantCalling(Process):
 
         if inputs.intervals:
             input_hc["intervals_bed"] = inputs.intervals
+            input_hc["advanced"][
+                "interval_padding"
+            ] = inputs.haplotype_caller.interval_padding
 
         hc = Data.create(
             process=BioProcess.get_latest(slug="vc-gatk4-hc"),
@@ -462,8 +564,8 @@ class WorkflowRnaseqVariantCalling(Process):
         input_filtration = {
             "vcf": hc,
             "ref_seq": inputs.ref_seq,
-            "filter_expressions": inputs.filter_expressions,
-            "filter_name": inputs.filter_name,
+            "filter_expressions": inputs.variant_filtration.filter_expressions,
+            "filter_name": inputs.variant_filtration.filter_name,
             "advanced": {
                 "cluster": 3,
                 "window": 35,
@@ -478,24 +580,51 @@ class WorkflowRnaseqVariantCalling(Process):
             name=f"Filtered genotypes ({inputs.reads.name})",
         )
 
-        if inputs.exclude_filtered:
-            input_select = {
-                "vcf": filtration,
-                "select_type": inputs.select_variants.select_type,
-                "exclude_filtered": True,
-                "advanced_options": {
-                    "java_gc_threads": inputs.advanced.java_gc_threads,
-                    "max_heap_size": inputs.advanced.max_heap_size,
-                },
-            }
+        snpeff_inputs = {
+            "variants": filtration,
+        }
 
-            Data.create(
-                process=BioProcess.get_latest(slug="gatk-select-variants-single"),
-                input=input_select,
-                name=f"Selected genotypes ({inputs.reads.name})",
-            )
+        if "GRCh37" in inputs.ref_seq.output.build:
+            snpeff_inputs["database"] = "GRCh37.75"
+        elif "GRCh38" in inputs.ref_seq.output.build:
+            snpeff_inputs["database"] = "GRCh38.99"
 
-        multiqc_tools = [inputs.reads, alignment, mark_duplicates, bqsr]
+        if inputs.snpeff.filtering_options:
+            snpeff_inputs["filtering_options"] = inputs.snpeff.filtering_options
+
+        if inputs.clinvar:
+            snpeff_inputs["dbsnp"] = inputs.clinvar
+
+        snpeff = Data.create(
+            process=BioProcess.get_latest(slug="snpeff-single"),
+            input=snpeff_inputs,
+            name=f"Annotated genotypes ({inputs.reads.name})",
+        )
+
+        mutations_inputs = {
+            "variants": snpeff,
+            "vcf_fields": inputs.mutations_table.vcf_fields,
+            "ann_fields": inputs.mutations_table.ann_fields,
+            "bam": preprocess,
+            "advanced": {
+                "split_alleles": inputs.mutations_table.split_alleles,
+                "show_filtered": inputs.mutations_table.show_filtered,
+                "gf_fields": inputs.mutations_table.gf_fields,
+            },
+        }
+
+        if inputs.mutations:
+            mutations_inputs["mutations"] = inputs.mutations
+        elif inputs.geneset:
+            mutations_inputs["geneset"] = inputs.geneset
+
+        Data.create(
+            process=BioProcess.get_latest(slug="mutations-table"),
+            input=mutations_inputs,
+            name=f"Mutations table ({inputs.reads.name})",
+        )
+
+        multiqc_tools = [inputs.reads, alignment]
 
         if inputs.preprocessing:
             multiqc_tools.append(preprocessing)